pyproject.toml
[project]
name = "fabricaHistologia_backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "lucasantonete@gmail.com", email = "lucasantonete@gmail.com"},
    {name = "Gabz047", email = "gabe1612mail@gmail.com"},
]
dependencies = [
    "django>=4.2.13",
    "djangorestframework>=3.15.2",
    "setuptools>=70.1.1",
    "drf-spectacular>=0.27.2",
    "passage-identity>=2.5.0",
    "openapi-client>=1.1.7",
    "django-cors-headers>=4.4.0",
    "google-api-python-client>=2.135.0",
    "openapi-python-client>=0.21.1",
    "dotenv>=0.0.5",
    "python-dotenv>=1.0.1",
    "Pillow>=10.4.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "django-extensions>=3.2.3",
    "graphviz>=0.20.3",
    "celery>=5.4.0",
    "django-celery-results>=2.5.1",
    "flower>=2.0.1",
    "amqp>=5.2.0",
    "pika>=1.3.2",
    "whitenoise>=6.7.0",
    "django-cloudinary-storage>=0.3.0",
    "django-filter>=24.3",
    "dj-database-url>=2.3.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python src/manage.py runserver localhost:8000"
runserver = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
migrate = "python src/manage.py migrate"
makemigrations = "python src/manage.py makemigrations"
celery = "celery -A django_project worker -l info"
populate = "python src/manage.py populate"


